Move app into src package layout and use uv
