+++ pyproject.toml
@@ -1,0 +1,17 @@
+[build-system]
+requires = ["uv_build==0.11.13"]
+build-backend = "uv_build"
+
+[project]
+name = "federated-data-management-portal"
+version = "1.1.0"
+description = "Dash dashboard for federated data management summaries."
+readme = "README.md"
+requires-python = ">=3.12,<3.13"
+license = { file = "LICENSE" }
+dependencies = [
+    "dash==4.1.0",
+    "dash-bootstrap-components==2.0.4",
+    "pandas==3.0.2",
+    "plotly==6.7.0",
+]
