+++ src/federated_data_management_portal/__init__.py
@@ -1,0 +1,1 @@
+"""Package marker for the federated data management portal distribution."""
